Add `.map()` and `.flatten()` to `Source`.
`map()` is a basic tool for cheap adaptations. It is not suitable for
arbitrary derived sources since it does no comparison, so spurious
notifications may pass through, and no memoization, so the mapped output
will be constructed anew for every `get()`. Still, it is useful for
adapting types, and for field projections where all fields are expected
to change together.

`flatten()` is less basic, but also something that is rather tricky to
write, so I hope that this implementation will be often useful.
I was hoping I could manage to do it without using any mutex, but I did
not succeed; notably, the previously-recorded inner source has to be
cloned, and we can’t express that without mutual exclusion. There could
perhaps be an implementation which doesn’t store the inner source at
all, but it would require adding lots of one-shot listeners, which has
its own costs and is contrary to the design of the library.