Introduce trait `FromListener<L, M>` partly replacing `IntoDynListener<M, L>`.
This is a breaking change.

This rearrangement of generic parameters allows `FromListener<L, M>` to
be implemented generically for a specific type outside the `nosy` crate,
such as an `enum` of listener types:

    enum Listener {
        Flag(nosy::FlagListener),
        ...
    }
    impl<M> nosy::FromListener<nosy::FlagListener, M> for Listener {

Previously, this would be a coherence error because `M` would have been
an “uncovered” type parameter occurring before `Listener`. The new
`Self` and parameter ordering makes it possible to do this because the
local type `Listener` appears first. By putting the destination type
first, we also

In exchange, we are losing the ability to write implementations of
the form

    impl<L> nosy::IntoDynListener<LocalMsg, L> for nosy::Flag {...}

but that implementation does not make sense unless the purpose of the
implementation is to implicitly adapt the message type `LocalMsg` to
another — which is not the purpose of implementing `IntoDynListener`.
I am confident that `M` should be last because it is the most likely to
remain generic except when the `Self` type is local.

The swap of `Self` and `L`, source and destination types, enables
blanket implementing `FromListener` for some destination type,
such as a new kind of type-erased smart pointer:

    impl<L, M> nosy::FromListener<L, M> for LocalListener {...}

In exchange, we lose the ability to implement conversion *from*
a local type to a generic one,

    impl<L> nosy::FromListener<LocalListener, ()> for L {...}

which could in principle be used for making a concrete listener fit
into more kinds of erased listener types (via an intermediate conversion
step). But I don’t think that’s more worthwhile.

In order to avoid regressing the convenience of accepting listeners,
the new sealed trait `IntoListener<L, M>` is blanket implemented
and serves the role of the old `IntoDynListener` trait.
(That does not make this not a breaking change, because the parameter
order has changed and `IntoListener` cannot be implemented.)