Add more `fmt::Pointer` implementations.
Unlike the existing implementations for `Cell` and `CellWithLocal`,
these do not include the type of the listener. I’m not sure which way
is better, but not doing so makes `Log`’s implementation possible in a
compositional way, without needing a `StoreLock::as_ptr()` method to get
the un-labeled address, which seems like a good sign to me.

All implementations are tested except for `Flatten`’s listeners.